requirements.txt
pydantic
requests
toml
unittest-xml-reporting
xmltodict
fabric
nuvla-api